+++ lib/transform/resolveNestedDefinitionTransformer.ts
@@ -11,9 +11,16 @@ export const resolveNestedDefinitionTransformer: SpecTransformer = {
   transform(spec, { jsonLoader, objSchemas, arrSchemas, primSchemas, allParams }) {
     const queue = new Array<string>();
 
-    const visitNestedDefinitions = (s: Schema | undefined, ref?: string) => {
+    const visitNestedDefinitions = (
+      s: Schema | undefined,
+      ref?: string,
+      isResponseSchema?: boolean
+    ) => {
       if (s === undefined || s === null || typeof s !== "object") {
         return;
+      }
+      if (isResponseSchema && s.type === "object" && s.properties === undefined) {
+        s.noPropertyWithTypeObject = true;
       }
       const schema = jsonLoader.resolveRefObj(s);
       if (visited.has(schema)) {
@@ -73,7 +80,7 @@ export const resolveNestedDefinitionTransformer: SpecTransformer = {
       onPath: visitParameters,
       onOperation: visitParameters,
       onResponse: (response) => {
-        visitNestedDefinitions(response.schema);
+        visitNestedDefinitions(response.schema, undefined, true);
       },
     });